add conversion function of one_hot on to_burn.rs modify one hot config

+++ crates/burn-import/src/burn/node/one_hot.rs
@@ -7,7 +7,7 @@ use quote::quote;
 
 #[derive(Config, Debug)]
 pub struct OneHotConfig {
-    pub depth: usize,
+    pub depth: i64,
     pub axis: isize,
     pub on_value: f32,
     pub off_value: f32,
